tests for useOutsideClick hook and fixed issue with useKeyPress test
